Let's Do a Readme
So I guess the purpose of this is to inform viewers what my project or Repository is all about, so to explain all I have to say is that it I don't really know what I am doing with it yet! This is for a class at UVU and I am learning all the basics. So yeah. That's what this is.